Remove legacy runtime typechecks
