+++ webpack/resources/reducer_support.ts
@@ -3,13 +3,11 @@ import {
   TaggedResource,
   ResourceName,
   SpecialStatus,
-  TaggedSequence
 } from "farmbot";
 import {
   sanityCheck,
   isTaggedResource,
 } from "./tagged_resources";
-import { generateUuid } from "./util";
 import {
   initialState as sequenceState,
   sequenceReducer as sequences,
@@ -33,17 +31,7 @@ import {
   initialState as helpState
 } from "../help/reducer";
 import { maybeTagSteps as dontTouchThis } from "./sequence_tagging";
-import {
-  recomputeLocalVarDeclaration
-} from "../sequences/step_tiles/tile_move_absolute/variables_support";
-
-const consumerReducer = combineReducers<RestResources["consumers"]>({
-  regimens,
-  sequences,
-  farm_designer,
-  farmware,
-  help
-} as any); // tslint:disable-line
+import { reindexResource } from "./reducer_indexing";
 
 export function emptyState(): RestResources {
   return {
@@ -91,6 +79,26 @@ export function emptyState(): RestResources {
 
 export const initialState: RestResources = emptyState();
 
+export function joinKindAndId(kind: ResourceName, id: number | undefined) {
+  return `${kind}.${id || 0}`;
+}
+
+/** Any reducer that uses TaggedResources (or UUIDs) must live within the
+ * resource reducer. Failure to do so can result in stale UUIDs, referential
+ * integrity issues and other bad stuff. The variable below contains all
+ * resource consuming reducers. */
+const consumerReducer = combineReducers<RestResources["consumers"]>({
+  regimens,
+  sequences,
+  farm_designer,
+  farmware,
+  help
+} as any); // tslint:disable-line
+
+/** The resource reducer must have the first say when a resource-related action
+ * fires off. Afterwards, sub-reducers are allowed to make sense of data
+ * changes. A common use case for sub-reducers is to listen for
+ * `DESTROY_RESOURCE_OK` and clean up stale UUIDs. */
 export const afterEach = (state: RestResources, a: ReduxAction<object>) => {
   state.consumers = consumerReducer({
     sequences: state.consumers.sequences,
@@ -105,62 +113,17 @@ export const afterEach = (state: RestResources, a: ReduxAction<object>) => {
 /** Helper method to change the `specialStatus` of a resource in the index */
 export const mutateSpecialStatus =
   (uuid: string, index: ResourceIndex, status = SpecialStatus.SAVED) => {
-    const resource = index.references[uuid];
-    if (resource) {
-      resource.specialStatus = status;
-    } else {
-      throw new Error("Refreshed a non-existent resource");
-    }
+    findByUuid(index, uuid).specialStatus = status;
   };
 
-export function addAllToIndex<T extends TaggedResource>(i: ResourceIndex,
-  kind: T["kind"],
-  all: T["body"][]) {
-  all.map(function (body) {
-    return addOneToIndex(i, kind, body, generateUuid(body.id, kind));
-  });
-}
-
-function addOneToIndex<T extends TaggedResource>(index: ResourceIndex,
-  kind: T["kind"],
-  body: T["body"],
-  uuid: string) {
-  const tr: TaggedResource = {
-    kind, body, uuid, specialStatus: SpecialStatus.SAVED
-    // tslint:disable-next-line:no-any
-  } as any;
+export function initResourceReducer(s: RestResources,
+  { payload }: ReduxAction<TaggedResource>): RestResources {
+  const tr = payload;
+  reindexResource(s.index, tr);
+  s.index.references[tr.uuid] = tr;
   sanityCheck(tr);
-  index.all.push(tr.uuid);
-  index.byKind[tr.kind].push(tr.uuid);
-  if (tr.body.id) { index.byKindAndId[tr.kind + "." + tr.body.id] = tr.uuid; }
   dontTouchThis(tr);
-  index.references[tr.uuid] = tr;
-}
-
-export function joinKindAndId(kind: ResourceName, id: number | undefined) {
-  return `${kind}.${id || 0}`;
-}
-
-const filterOutUuid =
-  (tr: TaggedResource) => (uuid: string) => uuid !== tr.uuid;
-
-export function removeFromIndex(index: ResourceIndex, tr: TaggedResource) {
-  const { kind } = tr;
-  const id = tr.body.id;
-  index.all = index.all.filter(filterOutUuid(tr));
-  if (index.byKind[tr.kind]) {
-    index.byKind[tr.kind] = index.byKind[tr.kind].filter(filterOutUuid(tr));
-    delete index.byKindAndId[joinKindAndId(kind, id)];
-    delete index.byKindAndId[joinKindAndId(kind, 0)];
-    delete index.references[tr.uuid];
-  } else {
-    console.log("No index found for tr.kind: " + tr.kind);
-  }
-}
-
-export function whoops(origin: string, kind: string): never {
-  const msg = `${origin}/${kind}: No handler written for this one yet.`;
-  throw new Error(msg);
+  return s;
 }
 
 export function findByUuid(index: ResourceIndex, uuid: string): TaggedResource {
@@ -172,29 +135,7 @@ export function findByUuid(index: ResourceIndex, uuid: string): TaggedResource {
   }
 }
 
-export function reindexResource(i: ResourceIndex, r: TaggedResource) {
-  removeFromIndex(i, r);
-  addOneToIndex(i, r.kind, r.body, r.uuid);
-}
-
-/** If the body of a sequence changes, we need to re-traverse the tree to pull
- * out relevant variable names. We try to avoid this via diffing. */
-export function maybeRecalculateLocalSequenceVariables(next: TaggedResource) {
-  (next.kind === "Sequence") && doRecalculateLocalSequenceVariables(next);
-}
-
-function doRecalculateLocalSequenceVariables(next: TaggedSequence) {
-  const recomputed = recomputeLocalVarDeclaration(next.body);
-  next.body.args = recomputed.args;
-  next.body.body = recomputed.body;
-}
-
-export function initResourceReducer(s: RestResources,
-  { payload }: ReduxAction<TaggedResource>): RestResources {
-  const tr = payload;
-  reindexResource(s.index, tr);
-  s.index.references[tr.uuid] = tr;
-  sanityCheck(tr);
-  dontTouchThis(tr);
-  return s;
+export function whoops(origin: string, kind: string): never {
+  const msg = `${origin}/${kind}: No handler written for this one yet.`;
+  throw new Error(msg);
 }
